vista del menú con backend
