+++ src/front/js/component/menu.js
@@ -1,5 +1,5 @@
 
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import andalogofood from "../../img/anda.png";
 import "../../styles/shoppingCart.css";
 
@@ -22,6 +22,9 @@ export const Menu = () => {
   const [spinner, setSpinner] = useState(false);
 
   const [showNotification, setShowNotification] = useState(false);
+
+  const [menus, setMenus] = useState([]);
+
 
   const mostrarNotificacion = () => {
     if (!showNotification) {
@@ -69,37 +72,53 @@ export const Menu = () => {
 
   };
 
+  useEffect(() => {
+    const fetchMenus = async () => {
+        try {
+            const response = await fetch(process.env.BACKEND_URL + 'api/menu');
+            const data = await response.json();
+            setMenus(data);
+        } catch (error) {
+            console.error("Error fetching menus:", error);
+        }
+    };
+
+    fetchMenus();
+}, []);
+
+
+
   const menuDay2 = {
-    LUNES: [
-      { id: "lunes_1", name: "Opción 1", description: "Arroz con trozos de pollo y verduras.", img: "https://cdn0.recetasgratis.net/es/posts/4/3/6/arroz_con_pollo_al_curry_28634_orig.jpg", price: "$350" },
-      { id: "Lunes_2", name: "Opción 2", description: "Ensalada de lechuga , tomate ,kiwi ,cebolla y rabanito", img: "https://www.nutrioli.com/wp-content/uploads/2016/06/Ensalada-de-lechugas-frutas-y-nuez-de-la-india-2.jpg", price: "$350" },
-      { id: "lunes_3", name: "Opción 3", description: "Albondigas de carne con salsa picante", img: "https://www.lacocinadelila.com/wp-content/uploads/2021/01/albondigas-de-pollo-600x450.jpg", price: "$350" },
-    ],
-    MARTES: [
-      { id: "Martes_1", name: "Opción 1", description: "Churrasco de pollo con ensalada de coliflor ,zanahoria y esparragos", img: "https://content.elmueble.com/medio/2024/10/03/pollo-a-la-plancha-con-verduras-al-pesto_a53bee68_241003155651_1200x1200.webp", price: "$350" },
-      { id: "Martes_2", name: "Opción 2", description: "Ensalda de coliflor ,repollo ,acelga ,zanahoria y queso rallado ", img: "https://upload.wikimedia.org/wikipedia/commons/2/23/Caesar_salad_%282%29.jpg", price: "$350" },
-      { id: "Martes_3", name: "Opción 3", description: "Panchos con panceta ahumada morron picadito y salsa ahumada", img: "https://sinreservas.com.ar/download/multimedia.normal.97a15526e9be66bf.53616e7a61727520283130295f6e6f726d616c2e77656270.webp", price: "$350" },
-    ],
-    MIÉRCOLES: [
-      { id: "Miércoles_1", name: "Opción 1", description: "Sopa crema de zapallo ,zanahoria, cebolla, morron y remolacha", img: "https://www.deliciosi.com/images/300/378/crema-de-calabaza.jpg", price: "$350" },
-      { id: "Miércoles_2", name: "Opción 2", description: "Cazuela de matambre con porotos y verudras", img: "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQnrVXtOErESITNnhn9MJMOptSFBKsEPA-9QA&s", price: "$300" },
-      { id: "Miércoles_3", name: "Opción 3", description: "Arroz primavera con choclo ,zanahoria ,arveja y tomate", img: "https://content.elmueble.com/medio/2024/09/06/arroz-con-verduras_ab2a54ac_240906144446_1200x1200.jpg", price: "$250" },
-    ],
-    JUEVES: [
-      { id: "Jueves_1", name: "Opción 1", description: "Milanesa de berenjena a la napolitana", img: "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcSeL6NmTbnojfQnaiMnYoLpiNqzOYnkNMBQHA&s", price: "$350" },
-      { id: "Jueves_2", name: "Opción 2", description: "Carne de cerdo con lechuga,zanahoria y rabanito", img: "https://truffle-assets.tastemadecontent.net/cdn-cgi/image/width=360/a2f94f01-742_friedchickensalad_square2.jpg", price: "$350" },
-      { id: "Jueves_3", name: "Opción 3", description: "Tallarines con tuco de carne picada", img: "https://www.frutamare.com/wp-content/uploads/2021/04/pasta-a-la-bolonesa.jpg", price: "$300" },
-    ],
-    VIERNES: [
-      { id: "Viernes_1", name: "Opción 1", description: "Ensalda de tomate , rucula y queso", img: "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcS7n0A8vIS8MRwnHMZhX-MxS2W_L2b0-6g8Jw&s", price: "$350" },
-      { id: "Viernes_2", name: "Opción 2", description: "Milanesa de pescado con papas fritas", img: "https://img.freepik.com/fotos-premium/weiner-schnitzel-empanado-casero-papas-fritas-pollo-frito-papas-fritas-estilo-comida-europea_1339-152591.jpg", price: "$250" },
-      { id: "Viernes_3", name: "Opción 3", description: "Sorrentinos de calabza con salsa 4 quesos", img: "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRcpQ3KilaPlZwatN49TzcFTRwbDXQ-VFWsx_tScOJkHuh7zyh8_rIVvc-r72wM6whCNfc&usqp=CAU", price: "$200" },
-    ],
-    SÁBADO: [
-      { id: "Sábado_1", name: "Opción 1", description: "Ensalda de garbanzos, tomate , brocoli y esparragos", img: "https://cdn0.recetasgratis.net/es/posts/8/9/0/ensalada_de_garbanzos_vegana_59098_600_square.jpg", price: "$350" },
-      { id: "Sábado_2", name: "Opción 2", description: "Pastel de canre con salsa blanca ", img: "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTPnXJcdnE2L04D4PAB0py1DthQJoctIfVfUw&s", price: "$300" },
-      { id: "Sábado_3", name: "Opción 3", description: "Brochet de verduras con carne de cerdo ", img: "https://i.pinimg.com/736x/bf/cf/ce/bfcfce36554e902259cbcad00169a51c.jpg", price: "$350" },
-    ]
+    // LUNES: [
+    //   { id: "lunes_1", name: "Opción 1", description: "Arroz con trozos de pollo y verduras.", img: "https://cdn0.recetasgratis.net/es/posts/4/3/6/arroz_con_pollo_al_curry_28634_orig.jpg", price: "$350" },
+    //   { id: "Lunes_2", name: "Opción 2", description: "Ensalada de lechuga , tomate ,kiwi ,cebolla y rabanito", img: "https://www.nutrioli.com/wp-content/uploads/2016/06/Ensalada-de-lechugas-frutas-y-nuez-de-la-india-2.jpg", price: "$350" },
+    //   { id: "lunes_3", name: "Opción 3", description: "Albondigas de carne con salsa picante", img: "https://www.lacocinadelila.com/wp-content/uploads/2021/01/albondigas-de-pollo-600x450.jpg", price: "$350" },
+    // ],
+    // MARTES: [
+    //   { id: "Martes_1", name: "Opción 1", description: "Churrasco de pollo con ensalada de coliflor ,zanahoria y esparragos", img: "https://content.elmueble.com/medio/2024/10/03/pollo-a-la-plancha-con-verduras-al-pesto_a53bee68_241003155651_1200x1200.webp", price: "$350" },
+    //   { id: "Martes_2", name: "Opción 2", description: "Ensalda de coliflor ,repollo ,acelga ,zanahoria y queso rallado ", img: "https://upload.wikimedia.org/wikipedia/commons/2/23/Caesar_salad_%282%29.jpg", price: "$350" },
+    //   { id: "Martes_3", name: "Opción 3", description: "Panchos con panceta ahumada morron picadito y salsa ahumada", img: "https://sinreservas.com.ar/download/multimedia.normal.97a15526e9be66bf.53616e7a61727520283130295f6e6f726d616c2e77656270.webp", price: "$350" },
+    // ],
+    // MIÉRCOLES: [
+    //   { id: "Miércoles_1", name: "Opción 1", description: "Sopa crema de zapallo ,zanahoria, cebolla, morron y remolacha", img: "https://www.deliciosi.com/images/300/378/crema-de-calabaza.jpg", price: "$350" },
+    //   { id: "Miércoles_2", name: "Opción 2", description: "Cazuela de matambre con porotos y verudras", img: "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQnrVXtOErESITNnhn9MJMOptSFBKsEPA-9QA&s", price: "$300" },
+    //   { id: "Miércoles_3", name: "Opción 3", description: "Arroz primavera con choclo ,zanahoria ,arveja y tomate", img: "https://content.elmueble.com/medio/2024/09/06/arroz-con-verduras_ab2a54ac_240906144446_1200x1200.jpg", price: "$250" },
+    // ],
+    // JUEVES: [
+    //   { id: "Jueves_1", name: "Opción 1", description: "Milanesa de berenjena a la napolitana", img: "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcSeL6NmTbnojfQnaiMnYoLpiNqzOYnkNMBQHA&s", price: "$350" },
+    //   { id: "Jueves_2", name: "Opción 2", description: "Carne de cerdo con lechuga,zanahoria y rabanito", img: "https://truffle-assets.tastemadecontent.net/cdn-cgi/image/width=360/a2f94f01-742_friedchickensalad_square2.jpg", price: "$350" },
+    //   { id: "Jueves_3", name: "Opción 3", description: "Tallarines con tuco de carne picada", img: "https://www.frutamare.com/wp-content/uploads/2021/04/pasta-a-la-bolonesa.jpg", price: "$300" },
+    // ],
+    // VIERNES: [
+    //   { id: "Viernes_1", name: "Opción 1", description: "Ensalda de tomate , rucula y queso", img: "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcS7n0A8vIS8MRwnHMZhX-MxS2W_L2b0-6g8Jw&s", price: "$350" },
+    //   { id: "Viernes_2", name: "Opción 2", description: "Milanesa de pescado con papas fritas", img: "https://img.freepik.com/fotos-premium/weiner-schnitzel-empanado-casero-papas-fritas-pollo-frito-papas-fritas-estilo-comida-europea_1339-152591.jpg", price: "$250" },
+    //   { id: "Viernes_3", name: "Opción 3", description: "Sorrentinos de calabza con salsa 4 quesos", img: "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRcpQ3KilaPlZwatN49TzcFTRwbDXQ-VFWsx_tScOJkHuh7zyh8_rIVvc-r72wM6whCNfc&usqp=CAU", price: "$200" },
+    // ],
+    // SÁBADO: [
+    //   { id: "Sábado_1", name: "Opción 1", description: "Ensalda de garbanzos, tomate , brocoli y esparragos", img: "https://cdn0.recetasgratis.net/es/posts/8/9/0/ensalada_de_garbanzos_vegana_59098_600_square.jpg", price: "$350" },
+    //   { id: "Sábado_2", name: "Opción 2", description: "Pastel de canre con salsa blanca ", img: "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTPnXJcdnE2L04D4PAB0py1DthQJoctIfVfUw&s", price: "$300" },
+    //   { id: "Sábado_3", name: "Opción 3", description: "Brochet de verduras con carne de cerdo ", img: "https://i.pinimg.com/736x/bf/cf/ce/bfcfce36554e902259cbcad00169a51c.jpg", price: "$350" },
+    // ]
   };
 
   const anotheroptions = [
@@ -282,106 +301,62 @@ export const Menu = () => {
     <MenuNavbar spinner={spinner}/>
 
     {/* Menu del dia */}
-    <div className="menudeldia2 mt-3" style={{ marginBottom: "20px", fontFamily: "Mulish, sans-serif" }}>
-      <div className="mb-5">
-        <h2 className="text-center" style={{ color: "rgb(56, 101, 229)" }}>MENÚ DE LA SEMANA</h2>
-        <div className="row">
-          {Object.keys(menuDay2).map((day) => (
-            <div key={day} className="col-12 mb-4">
-              <div
-                style={{
-                  backgroundColor: "rgba(56, 101, 229, 0.5)",
-                  borderRadius: "10px",
-                  padding: "15px",
-
-                }}
-              >
-                <h3 className="text-center" style={{ color: "white" }}>{day}</h3>
-                <div className="row">
-                  {menuDay2[day].map((item) => (
-                    <div key={item.id} className="col-12 col-md-6 col-lg-4 mb-3">
-                      <div
-                        className="card mx-auto"
-                        style={{
-                          borderRadius: "10px",
-                          overflow: "hidden",
-                          width: "200px",
-                          height: "200px",
-                          border: "none",
-                          boxShadow: "0 4px 8px rgba(0, 0, 0, 0.2)",
-                        }}
-                      >
-                        <div
-                          className="position-relative"
-                          style={{
-                            backgroundColor: "rgb(56, 101, 229)",
-                            color: "white",
-                            fontSize: "14px",
-                            fontWeight: "bold",
-                            textAlign: "center",
-                          }}
-                        >
-                          {item.name}
-                        </div>
-                        <img
-                          src={item.img}
-                          className="card-img-top img-fluid"
-                          alt={item.name}
-                          style={{
-                            objectFit: "cover",
-                            height: "100px",
-                            width: "100%",
-                          }}
-                        />
-                        <div
-                          style={{
-                            fontWeight: "bold",
-                            fontSize: "14px",
-                            color: "rgb(56, 101, 229)",
-                            textAlign: "center",
-                            height: "120px",
-                          }}
-                        >
-                          {item.price}
-                        </div>
-
-                        <div className="card-body text-center p-2">
-                          <div className="d-flex justify-content-between">
-
-                            <button
-                              className="btn " onClick={() => { handleClick(item); handleNotificacion() }}
-                              style={{
-                                backgroundColor: "rgb(56, 101, 229)",
-                                color: "white",
-                                fontSize: "12px",
-                                borderRadius: "10px",
-                              }}
-                            >
-                              Añadir al Carrito
-                            </button>
-                            <button
-                              className="btn"
-                              style={{
-                                backgroundColor: "transparent",
-                                fontSize: "20px",
-                              }}
-                              onClick={() => irAFeedback(item)}
-                            >
-                              ⭐
-                            </button>
-                          </div>
-                        </div>
-
-                      </div>
-                    </div>
-                  ))}
-                </div>
+    <div className="container">
+  <h2>MENÚ DE LA SEMANA </h2>
+  <div className="row">
+    {menus.map((menu) => (
+      <div key={menu.id} className="col-md-4">
+        <div className="card">
+          <img
+            src={menu.img}
+            alt={menu.name}
+            className="card-img-top"
+          />
+          <div className="card-body">
+            <h5 className="card-title">{menu.name}</h5>
+            <p className="card-text">{menu.description}</p>
+            <p className="card-text">
+              <strong>Precio:</strong> {menu.price}
+            </p>
+            <p className="card-text">
+              <strong>Día:</strong> {menu.day}
+            </p>
+            <div className="card-body text-center p-2">
+              <div className="d-flex justify-content-between">
+                <button
+                  className="btn"
+                  onClick={() => {
+                    handleClick(menu);
+                    handleNotificacion();
+                  }}
+                  style={{
+                    backgroundColor: "rgb(56, 101, 229)",
+                    color: "white",
+                    fontSize: "12px",
+                    borderRadius: "10px",
+                  }}
+                >
+                  Añadir al Carrito
+                </button>
+                <button
+                  className="btn"
+                  style={{
+                    backgroundColor: "transparent",
+                    fontSize: "20px",
+                  }}
+                  onClick={() => irAFeedback(menu)}
+                >
+                  ⭐
+                </button>
               </div>
             </div>
-          ))}
+          </div>
         </div>
       </div>
-    </div>
+    ))}
+  </div>
+</div>
+
 
     {/* Notificación */}
